requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
mysql-connector-python==8.1.0